requirements.txt
python-pptx>=0.6.23
certifi>=2024.0.0
lxml>=4.9.0
pymupdf>=1.23.0
